Move console classes from _pyodide to pyodide
